+++ packages/backend-function/src/runtime/index.ts
@@ -1,1 +1,10 @@
-export { getAmplifyClientsConfiguration } from './get_amplify_clients_configuration.js';
+export {
+  getAmplifyDataClientConfig,
+  DataClientConfig,
+  DataClientEnv,
+  DataClientError,
+  DataClientReturn,
+  InvalidConfig,
+  LibraryOptions,
+  ResourceConfig,
+} from './get_amplify_clients_configuration.js';