+++ packages/integration-tests/src/test-projects/data-and-function/amplify/backend.ts
@@ -1,5 +1,6 @@
 import { defineBackend } from '@aws-amplify/backend';
 import { data } from './data/resource.js';
 import { todoCount } from './functions/todo-count/resource.js';
+import { noopImport } from './functions/noop-import/resource.js';
 
-const backend = defineBackend({ data, todoCount });
+const backend = defineBackend({ data, todoCount, noopImport });