+++ packages/integration-tests/src/test-projects/data-and-function/amplify/functions/noop-import/handler.ts
@@ -4,7 +4,8 @@ import { generateClient } from 'aws-amplify/data';
 import type { Schema } from '../../data/resource.js';
 import { getAmplifyDataClientConfig } from '@aws-amplify/backend/function/runtime';
 // @ts-ignore
-import { env } from '$amplify/env/todo-count.js';
+import { env } from '$amplify/env/noop-import.js';
+import { S3Client } from '@aws-sdk/client-s3';
 
 const { resourceConfig, libraryOptions } = await getAmplifyDataClientConfig(
   env
@@ -15,6 +16,7 @@ Amplify.configure(resourceConfig, libraryOptions);
 const client = generateClient<Schema>();
 
 export const handler: Handler = async () => {
-  const todos = await client.models.Todo.list();
-  return todos.data.length;
+  const _s3Client = new S3Client();
+  const _todos = await client.models.Todo.list();
+  return 'STATIC TEST RESPONSE';
 };