+++ packages/integration-tests/src/test-projects/data-and-function/amplify/functions/noop-import/resource.ts
@@ -1,0 +1,7 @@
+import { defineFunction } from '@aws-amplify/backend';
+
+export const noopImport = defineFunction({
+  name: 'noop-import',
+  entry: './handler.ts',
+  timeoutSeconds: 30,
+});