+++ packages/integration-tests/src/test-projects/data-and-function/test-types/env/todo-count.ts
@@ -1,0 +1,18 @@
+/* eslint-disable unicorn/filename-case */
+/* eslint-disable check-file/filename-naming-convention */
+// we have to match the naming convention of the generated env shim file
+
+// we have to match the expected environment variable names
+/* eslint-disable @typescript-eslint/naming-convention */
+export const env = process.env as {
+  TEST_NAME_BUCKET_NAME: string;
+  AWS_REGION: string;
+  AWS_ACCESS_KEY_ID: string;
+  AWS_SECRET_ACCESS_KEY: string;
+  AWS_SESSION_TOKEN: string;
+  TEST_SECRET: string;
+  TEST_SHARED_SECRET: string;
+  AMPLIFY_DATA_GRAPHQL_ENDPOINT: string;
+  AMPLIFY_DATA_MODEL_INTROSPECTION_SCHEMA_BUCKET_NAME: string;
+  AMPLIFY_DATA_MODEL_INTROSPECTION_SCHEMA_KEY: string;
+};